Fix duplicate instance bug with parsers-0.1.0 and parsec-parsers.
The parsers package as of 0.1.0 includes the instances that were previously
contained in parsec-parsers, making the latter unnecessary. This should fix the
duplicate instance error that arises if both are installed alongside each other.

This change will require forced update of the parsers package to version 0.1.0
if a previous version is present.