Begin changes to allow Dataspaces to be in any Monad
The Monad must contain the EngineM Value monad, but this isn't
enforced yet.  The idea is that the DataSpace class knows how to lift
from Engine back to its monad, and can pass that function into
things like foldFile, etc.  This doesn't work yet, so liftM is just passed
around and I use liftEngine for now.