Testing Endpoint doWrite on NetworkHandle
Changed ClientTest.cpp to wrap the NetworkHandle with an Endpoint for
testing.

Basic reading/writing over the wire is now working with an Endpoint.
However there is more work to be done here, as a partial message or even
multiple messages may be read with a single call to async_read(). The
receiver must figure out how to construct proper message(s) from the socket
data. Yanif and I discussed potentially prefixing each message with a
header to indicate the size of the message in bytes.

Also added some more verbose logging upon incoming socket data which is
useful for debugging now, but will probably become annoying as we get
further into Engine development.