Fix non-alphabetical record issues.
There's a subtle issue when creating literal records -- we might store
record fields in alphabetic order, but if the programmer has
side-effects in their field initializers, we should probably respect the
order of those effects.

Therefore, only the values of the record fields are shuffled around, not
the effects derived from their creation.