Switch lambdas to const&, turn offending binds to projections.
A bind can't be used if the thing its destructing is a const&, since
writeback would be illegal. Normally we'd transform something to const&
only if it was immutable, but...

It looks like we need to split bind's functionality into two --
destructuring and writeback. The first can be done on immutable values,
the second not so much.