Fix groupBy segmentation faults across the board.
DON'T MAKE EVERYTHING `const&'!.

It isn't a safe blanket transformation, and it's damn near impossible to
debug when it goes wrong.

Long story short -- λ-capture in C++ bundles the closure with the
function-object, and the mutability of captured values (indeed, anything
at all about th captured values) isn't part of the type of that object.

If you pass a function-object that has reference captured values into a
another function which asks for a constant reference to a function of
the same type, and the function you're passing in tries to modify some
of its reference captured values, well, the typechecker isn't going to
help you.

Moral of the story: Don't accept functions by constant reference, unless
you know for certain that they will never capture anything by reference.
And since you can't ever know that for certain, just don't do it at all.

This might tie for "most insidious bug I've fixed", with ae1f4b6.