runtime: added serialization for shared_ptr
We serialize by getting the pointer out of the shared_ptr. This means
that no notion of sharing is kept -- if you serialize a shared_ptr once,
and then serialize the same shared_ptr again, it'll create 2 different
structures when deserialized. This should be fine for our purposes (at
least for now). There's no way to get at the sharing information, even
if we convert to boost::shared_ptr locally, so the only way to keep sharing info
(if we want it) would be by using boost::shared_ptr instead of std::shared_ptr