Handle interaction between case/of and return.
This interaction is extremely subtle and specific. If:

1. The case/of is the last expression in a function (its result is the
   one being returned), and
2. The option value being case-d on is a global, and
3. The global in question is read/written inside the case branches,
   independent of the substructure bound within, then

write-back will not happen in this case (the write-back statements are
incorrectly stashed after the return statement). Correctly handling this
case requires the materialization engine to send yet another piece of
information to the code generator, which cannot overlap with the
existing decision mechanism. I'm choosing not to put this in now, the
above conditions 2 and 3 are unlikely individually, together they almost
never happen.