Revert to move-based loop accumulator in fold generation.
The previous method for eliminating the copy of a loop accumulator by
using the passed-down name falls flat in the face of expressions such
as `x = x.fold(...)`.

This patch reverts the copy elision changes introduced earlier, and
switches the loop accumulator to be reified by move, which is _almost_
as efficient.